updating package to latest npm conventions. This should allow the user to do a simple require('grunt-npm-helper") instead of requiring the absolute path.
